+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/entities/station/Station.kt
@@ -1,9 +1,9 @@
-package fr.gouv.cnsp.monitorfish.domain.entities.control_unit
+package fr.gouv.cnsp.monitorfish.domain.entities.station
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ControlUnitStation(
+data class Station(
     val id: Int,
     val latitude: Double,
     val longitude: Double,
